Made dimensions 32bit

+++ IWorldObject.h
@@ -11,11 +11,11 @@ namespace alt
 	{
 	public:
 		virtual Position GetPosition() const = 0;
-		virtual int16_t GetDimension() const = 0;
+		virtual int32_t GetDimension() const = 0;
 
 #ifdef ALT_SERVER_API
 		virtual void SetPosition(Position pos) = 0;
-		virtual void SetDimension(int16_t dimension) = 0;
+		virtual void SetDimension(int32_t dimension) = 0;
 #endif // ALT_SERVER_API
 
 	protected:
